refactor(chrome-extension): extract preview into a separate module
