+++ apps/chrome-extension/src/features/textAdaptationPreview/index.ts
@@ -1,0 +1,3 @@
+export { default as TextAdaptationPreview } from './ui/TextAdaptationPreview.vue'
+// TODO: encapsulate the usage of the function inside the module
+export { adaptHtmlElementAsyncFn } from './lib/visualEngine/adaptHtmlElementAsync'
