+++ apps/chrome-extension/src/features/textAdaptationPreview/lib/visualEngine/adaptHtmlElementAsync.ts
