+++ apps/chrome-extension/src/features/textProfilesDropdown/index.ts
