+++ apps/chrome-extension/src/features/textProfilesDropdown/ui/TextProfilesDropdown.vue
